+++ packages/client/src/client.ts
@@ -1,5 +1,6 @@
 import { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 import {
+  BlockInfo,
   ChainHead$,
   getObservableClient,
   withArchive,
@@ -12,6 +13,7 @@ import {
 import {
   Observable,
   catchError,
+  combineLatest,
   defer,
   firstValueFrom,
   from,
@@ -206,6 +208,27 @@ export function createClient(
   })
   const { getChainSpecData } = rawClient
 
+  const withArchiveBlock =
+    <T extends any[], R>(
+      fn: (blockHash: string, ...args: T) => Observable<R>,
+    ) =>
+    (
+      blockHash: string,
+      ...args: T
+    ): Observable<{ value: R; block: BlockInfo }> =>
+      combineLatest([
+        fn(blockHash, ...args),
+        archive.header$(blockHash).pipe(
+          map(({ number, parentHash }) => ({
+            hash: blockHash,
+            number,
+            parent: parentHash,
+            // Not correct, but it would be overkill to fix. Only relevant on chainHead, not on archive.
+            hasNewRuntime: false,
+          })),
+        ),
+      ]).pipe(map(([value, block]) => ({ value, block })))
+
   const { genesis$, ..._chainHead } = client.chainHead$()
   const archive = client.archive(_chainHead.getRuntime$)
   const chainHead: ChainHead$ = {
@@ -215,11 +238,17 @@ export function createClient(
       catchError(() => genesis$),
       shareReplay(1),
     ),
-    storage$: withArchive(_chainHead.storage$, archive.storage$),
+    storage$: withArchive(
+      _chainHead.storage$,
+      withArchiveBlock(archive.storage$),
+    ),
     body$: withArchive(_chainHead.body$, archive.body$),
     call$: withArchive(_chainHead.call$, archive.call$),
     header$: withArchive(_chainHead.header$, archive.header$),
-    eventsAt$: withArchive(_chainHead.eventsAt$, archive.eventsAt$),
+    eventsAt$: withArchive(
+      _chainHead.eventsAt$,
+      withArchiveBlock(archive.eventsAt$),
+    ),
     storageQueries$: withArchive(
       _chainHead.storageQueries$,
       archive.storageQueries$,
@@ -287,7 +316,7 @@ export function createClient(
         storage$(at ?? null, "value", () => {
           const hex = key.match(HEX_REGEX)?.[1]
           return hex ? `0x${hex}` : Binary.fromText(key).asHex()
-        }),
+        }).pipe(map((v) => v.value)),
         signal,
       ),