+++ packages/observable-client/tests/mockSubstrateClient.ts
@@ -1,3 +1,4 @@
+import { BlockInfo } from "@/chainHead"
 import {
   ChainHead,
   FollowEventWithRuntime,